refactor: move utils.ts to `common` folder
